+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Version [3.4.0](https://github.com/cedx/core.js/compare/v3.3.0...v3.4.0)
 - Renamed the `Async` module to `Threading`.
 - Renamed the `Cache` and `CacheSerializer` interfaces to `ICache` and `ICacheSerializer`.
-- Renamed the `Task.value` peroperty to `result`.
+- Renamed the `Region` class to `RegionInfo`.
+- Renamed the `Task.value` property to `result`.
 - Renamed the values of the `TaskStatus` enumeration.
 - Added the `ICache.getOrCreate()` method.
 - Added the `isCompleted`, `isCompletedSuccessfully` and `isFaulted` properties to the `Task` class.